% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_vol.R
\name{eff_vol}
\alias{eff_vol}
\title{Calculate the effective volume of a respirometer}
\usage{
eff_vol(resp_vol = NULL, spec_vol = NULL, spec_mass = NULL,
  spec_density = NULL, t = NULL, S = NULL, P = 1.013253)
}
\arguments{
\item{resp_vol}{numeric (L). Volume of the respirometer or loop.}

\item{spec_vol}{numeric (L). Displacement volume of the specimen. Enter
either this or \code{spec_mass}.}

\item{spec_mass}{numeric (kg). Mass of the specimen. Enter either this or
\code{spec_vol}.}

\item{spec_density}{numeric (kg/m^3). Density of the specimen. Optional. Used to convert \code{spec_mass} to a volume.}

\item{t}{numeric (°C). Water temperature. Optional. Used to calculate density if
\code{spec_density} is left NULL.}

\item{S}{numeric (ppt). Water salinity. Optional. Used to calculate density if
\code{spec_density} is left NULL.}

\item{P}{numeric (bar). Atmospheric pressure. Optional. Used to calculate density if
\code{spec_density} is left NULL. Defaults to 1.013253.}
}
\description{
Calculate the effective volume of a respirometer.
}
\details{
Adjusts the known water volume of an aquatic respirometer (or
respirometer loop) for the volume of the specimen. This can be entered
either as a straight displacement volume of the specimen for simple
subtraction (\code{spec_vol}), or as a mass (\code{spec_mass}), in which case the
\code{spec_density} is required to calculate the displacement volume. The
\code{spec_density} can be entered as a value in kg/m^3. Alternatively, it can
be assumed the specimen is the same density as the water (this is the case
if it is neutrally buoyant). This can be entered as the \code{spec_density}
value, for example typical seawater density is around 1026 kg/m^3.
Alternatively it can be calculated internally, in which case temperature
(\code{t}) and salinity (\code{S}) are required, along with atmospheric pressure
(\code{P}). Note, that since \code{P} has a negligible effect on water density within
the typical ranges experienced at sea level, the default of 1.013253 bar is
used unless otherwise specified.
}
\section{Inputs}{
 Inputs must be in SI units:
\itemize{
\item \code{resp_vol} and \code{spec_vol} in \code{L}
\item \code{spec_mass} in \code{kg}
\item \code{spec_density} in \code{kg/m^3}
\item \code{t} in °C
\item \code{S} in ppt
\item \code{P} in bar (Defaults to 1.013253)
}
}

\section{Output}{
 Output is a single numeric value for the effective volume in
L.
}

\examples{
## Simple displacement volume subtraction.
eff_vol(resp_vol = 1, spec_vol = 0.1)

## Mass & density to volume subtraction.
eff_vol(resp_vol = 1, spec_mass = 0.2, spec_density = 1026)

## Mass, with t, S used to calculate water density
eff_vol(resp_vol = 1, spec_mass = 0.2, t = 10, S = 35)

}
\seealso{
spec_density
}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
