% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush_time.R
\name{flush_time}
\alias{flush_time}
\title{Flush time}
\usage{
flush_time(volume, flow, tolerance)
}
\arguments{
\item{volume}{numeric. Volume of the respirometer in L.}

\item{flow}{numeric. Flow rate of the input of new water in L/s.}

\item{tolerance}{numeric. Proportion of the \code{volume} to be exchanged. See
Details.}
}
\value{
Time in seconds for the \code{tolerance} proportion of the \code{volume} to be
exchanged.
}
\description{
Calculates the time taken to flush a respirometer with new
water.
}
\details{
\code{flush_time} calculates the time taken to flush a respirometer with
new water. It requires the volume of the respirometer (in L) and the flow
rate (in L/s) of the input of new water.

The function calculates the time in seconds to exchange the proportion of
water described by the \code{tolerance} input, which by default is \code{0.99}. That
is, the function calculates the time taken to exchange 99\% of the water.
The \code{tolerance} is required because of \emph{mixing} during the exchange.
Because old and new water are always mixing, the exchange of water is
asymptotic: you can never replace 100\% of the old water; there will be an
ever decreasing proportion of it remaining (down to the final molecule,
assuming perfect mixing). For all practical purposes, a \code{tolerance} value
of \code{0.99} is sufficient. It is also important to note that the function
takes no account of any shared reservoir of water, that is when the 'old'
water is flushed out into the source of the 'new' water.
}
\examples{
## How long to exchange 99\% of a 10L respirometer when the input pump has
## a flow rate of 0.15 L/s
flush_time(volume = 10, flow = 0.15, tolerance = 0.99)

}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
