% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ten.R
\name{q_ten}
\alias{q_ten}
\title{Calculates the Q10 temperature relationship}
\usage{
q_ten(rate1 = NULL, rate2 = NULL, t1 = NULL, t2 = NULL, q10 = NULL)
}
\arguments{
\item{rate1}{numeric. The rate at temperature 1 (\code{t1}).}

\item{rate2}{numeric. The rate at temperature 2 (\code{t2}).}

\item{t1}{numeric (°C). The \emph{lower} temperature at which \code{rate1} was
determined.}

\item{t2}{numeric (°C). The \emph{upper} temperature at which \code{rate2} was
determined.}

\item{q10}{numeric (°C). The Q10, the ratio by which a rate changes with a
10°C increase in temperature.}
}
\description{
The \code{q_ten} function calculates any of the five parameters in
the Q10 temperature relationship for physiological or chemical processes.
Q10 describes the ratio by which a physiological or chemical rate changes
with a 10°C increase in temperature.
}
\details{
The function has five inputs, of which four should be entered and
one left NULL. The NULL input is the parameter to be calculated. For
example, two rates (\code{rate1}, \code{rate2}) determined at two temperatures (\code{t1},
\code{t2}) can be used to calculate the \code{q10}. Alternatively, a \code{q10} and a rate
at a particular temperature can be used to predict a rate at a different
temperature. See examples.

See also the function \code{Q10} in the
\href{https://cran.r-project.org/web/packages/respirometry/index.html}{\code{respirometry}}
package by Matthew Birk, which has additional functionality.
}
\examples{
## Determine Q10 from two rates at two known temperatures
q_ten(rate1 = 2, rate2 = 3, t1 = 10, t2 = 15, q10 = NULL)

## Predict a rate at a higher temperature using a known Q10
q_ten(rate1 = 2, rate2 = NULL, t1 = 10, t2 = 15, q10 = 2.25)

## Predict a rate at a lower temperature using a known Q10
q_ten(rate1 = NULL, rate2 = 3, t1 = 10, t2 = 15, q10 = 2.25)

## Predict the lower temperature at which rate1 will occur
## using a known Q10
q_ten(rate1 = 2, rate2 = 3, t1 = NULL, t2 = 15, q10 = 2.25)

## Predict the upper temperature at which rate2 will occur
## using a known Q10
q_ten(rate1 = 2, rate2 = 3, t1 = 10, t2 = NULL, q10 = 2.25)

}
\seealso{
respirometry::Q10
}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
