% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_density.R
\name{spec_density}
\alias{spec_density}
\title{Calculate specimen density}
\usage{
spec_density(wet_mass = NULL, buoy_mass = NULL, t = NULL, S = NULL, P
  = 1.013253)
}
\arguments{
\item{wet_mass}{numeric (kg). Wet mass of specimen.}

\item{buoy_mass}{numeric (kg). Buoyant mass of specimen.}

\item{t}{numeric (°C). Water temperature at which \code{buoy_mass} was determined.}

\item{S}{numeric (ppt). Water salinity at which \code{buoy_mass} was determined.}

\item{P}{numeric (bar). Atmospheric pressure at which \code{buoy_mass} was
determined. Defaults to 1.013253.}
}
\description{
Given a wet mass (\code{wet_mass}) and buoyant mass (\code{buoy_mass}),
this function calculates the density of the specimen. For species such as
molluscs this can be used to determine the ratio of shell to tissue mass.
It can also be used to calculate specimen volume for correcting the water
volume of a respirometer (see \code{\link{eff_vol}}). The temperature
(\code{t}) and salinity (\code{S}) at which the buoyant mass was determined
are required (as is atmospheric pressure (\code{P}) which defaults to 1.013253
bar, if not otherwise specified).
}
\section{Inputs}{
 Inputs must be in SI units:
\itemize{
\item \code{wet_mass} and \code{buoy_mass} in \code{kg}
\item \code{t} in °C
\item \code{S} in ppt
\item \code{P} in bar (Defaults to 1.013253)
}
}

\section{Positively buoyant specimens}{
 The function \emph{should} work with
positively buoyant specimens (i.e. less dense than water). Given a bouyant
weight measured as buoyancy (i.e. lift force on a balance while in water),
this can be entered as a negative value for \code{buoy_mass} and the function
should return the correct density value. However, none of this has been
tested.
}

\section{Output}{
 Output is a single numeric value for the density in kg/m^3.
}

\seealso{
eff_vol
}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
