% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_rate.R
\name{split_rate}
\alias{split_rate}
\title{Divide a metabolic rate amongst a group}
\usage{
split_rate(..., tR = NULL, b = 0.75)
}
\arguments{
\item{tR}{numeric. Total group metabolic rate}

\item{b}{numeric. Metabolic scaling exponent}

\item{...}{numeric. A vector of body masses of all individuals in group}
}
\value{
A \code{list} object containing 5 numeric elements:

$a = \code{a} in the rate-mass power equation. Determined by the function

$b = \code{b} in the rate-mass power equation. User entered

$tR = total group rate. User entered

$masses = specimen masses. User entered.

$indiv.rates = Primary output of interest. Group rate divided between
individuals. Sum should therefore equal tR. Determined by the function
}
\description{
Divide a metabolic rate amongst a group of individuals.
}
\details{
Divides a group metabolic rate amongst individuals, given body masses of
each and a scaling exponent.
}
\examples{
split_rate(c(2, 3, 4, 5, 6), tR = 500, b = 0.75)

}
\author{
Nicholas Carey - \link{nicholascarey@gmail.com}
}
