% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_to_vol.R
\name{wm_to_vol}
\alias{wm_to_vol}
\title{Convert a mass of water to a volume}
\usage{
wm_to_vol(mass, t, S, P)
}
\arguments{
\item{mass}{numeric (kg). Water mass to convert to volume}

\item{t}{numeric (°C). Temperature}

\item{S}{numeric (ppt). Salinity}

\item{P}{numeric (bar). Atmospheric pressure. Defaults to 1.013253}
}
\description{
\code{wm_to_vol} converts a water mass to a water volume.
Requires temperature, salinity and atmospheric pressure. These are used to
calculate the density, which is then used to convert the mass to volume.
}
\section{Inputs}{
 Inputs must be in SI units:
\itemize{
\item \code{mass} in \code{kg}
\item \code{t} in °C
\item \code{S} in ppt
\item \code{P} in bar (Defaults to 1.013253)
}
}

\section{Output}{
 Output is a single numeric value for the water volume in L.
}

\examples{
## Should return 0.9767414
wm_to_vol(mass = 1, t = 23, S = 35)

}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
