% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_rate.R
\name{split_rate}
\alias{split_rate}
\title{Divide a metabolic rate amongst a group}
\usage{
split_rate(tR = NULL, masses, b = 0.75, units = NULL)
}
\arguments{
\item{tR}{numeric. Total group metabolic rate}

\item{masses}{numeric. A vector of body masses of all individuals in group}

\item{b}{numeric. Metabolic scaling exponent}

\item{units}{string. Units of the rate. Extracted from \code{convert_rate} object or
can be entered by the user. For information only, does not affect any
calculations.}
}
\description{
Divide a metabolic rate amongst a group of individuals.
}
\details{
Divides a group metabolic rate amongst individuals, given body
masses of each and a scaling exponent.

Take care to enter the correct scaling exponent (\code{b}). This is usually
(with exceptions) a positive value between 0.66 and 1. If your \code{b} value is
less than this, especially if it is less than 0.33, and \emph{especially} if it
is negative, then it is likely a \emph{mass-specific} scaling exponent. The
correct scaling exponent is the positive difference of this value from 1.
For example, for a scaling exponent of 0.75, the mass-specific scaling
exponent would be -0.25.

If, for whatever reason, you want to do a simple per-capita division of the
rate regardless of the body masses, enter \code{b = 0}. Or just divide it by the
number of individuals.

Units can be entered, e.g. \code{units = "mg/h"}, if the user wants to save
these in the output for reference, however they do not affect calculations
in any way. If a \code{respR::convert_rate} object is used for the rate (see
below), the units are extracted from this and any other \code{units} input
ignored.
}
\section{Sign of the rate}{
 NOTE: both negative and positive rates can be
entered. In respirometry experiments, rates are typically reported as
positive values. These can be entered as is. In the case of
\code{respR::convert_rate} objects, extracted rates will typically be
\emph{negative}, and these are left unchanged in the \code{split_rate} function. In
\code{respR}, to be mathematically consistent (since they represent oxygen
depletion), respiration rates are represented by negative slopes, and
therefore rates returned as negative. You can enter the rate as the usually
reported postive value, or as a negative: the function will work with
either. Returned rates and intercept \code{a} will be identical in value, except
for the sign. In effect, if you enter a negative rate you can simply ignore
the signs for \code{a} and individual rates in the output.
}

\section{\code{respR} integration}{
 For total rate (\code{tR}) the function accepts
objects saved from the \code{respR}
(\url{https://github.com/januarharianto/respR}) \code{convert_rate} function. In
this case, the rate and units are automatically extracted. However, if it
contains a mass-specific rate (i.e. the rate has been adjusted in
\code{respR::convert_rate} to a specific mass), no conversion is done and a
warning is returned. Only absolute, that is non-mass specific, respiration
rates should be divided in this way.
}

\section{Output}{
 Output is a \code{list} object containing 7 elements:

\verb{$a} = \code{a}, the intercept in the mass~rate power equation. Determined by
the function.

\verb{$b} = \code{b}, the exponent in the mass~rate power equation. User entered.

\verb{$tR} = total group rate. User entered.

\verb{$masses} = specimen masses. User entered.

\verb{$units} = units of the rate. User entered or extracted from \code{convert_rate}
object. For information only, does not affect any calculations.

\verb{$indiv.rates} = Primary output of interest. Group rate divided between
individuals. Sum should therefore equal tR. Determined by the function

\verb{$input} = origin of \code{tR} and \code{units}. Either \code{manual} entry or
\code{convert_rate} object
}

\examples{
split_rate(tR = 500, masses = c(2, 3, 4, 5, 6), b = 0.75)

}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
